Move response functions to response utils
